src/test/java/org/identifiers/cloud/hq/ws/registry/models/validators/PrefixRegistrationRequestValidatorInstitutionNameTest.java
package org.identifiers.cloud.hq.ws.registry.models.validators;

import org.identifiers.cloud.hq.ws.registry.api.requests.ServiceRequestRegisterPrefixPayload;
import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.ExpectedException;
import org.junit.runner.RunWith;
import org.junit.runners.JUnit4;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.function.Function;

import static org.junit.Assert.*;

@SpringBootTest
@RunWith(JUnit4.class)
public class PrefixRegistrationRequestValidatorInstitutionNameTest {

    PrefixRegistrationRequestValidatorInstitutionName validator =
            new PrefixRegistrationRequestValidatorInstitutionName();

    Function<String, ServiceRequestRegisterPrefixPayload> fnc =
            (p) -> new ServiceRequestRegisterPrefixPayload().setInstitutionName(p);

    @Rule
    public ExpectedException exceptionRule = ExpectedException.none();

    @Test
    public void validateNullInstitutionName() {
        exceptionRule.expect(PrefixRegistrationRequestValidatorException.class);
        validator.validate(fnc.apply(null));
    }

    @Test
    public void validateBlankInstitutionName() {
        exceptionRule.expect(PrefixRegistrationRequestValidatorException.class);
        validator.validate(fnc.apply(""));
    }

    @Test
    public void validateValidInstitutionName() {
        assertTrue(validator.validate(fnc.apply("EBI")));
    }
}